feat: update migrations

+++ migrations/20240718130704_create_plan_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE plans (
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   name VARCHAR(255) NOT NULL,
   description TEXT NOT NULL,
-  price DECIMAL(10, 2) NOT NULL,
+  price BIGINT NOT NULL,
   is_active BOOLEAN DEFAULT TRUE,
   tags TEXT[] DEFAULT '{}',
   trial_days INT DEFAULT 0,
